Updated components/__init__.py
